pyproject.toml
[tool.poetry]
name = "edgewalker"
version = "0.0.1"
description = "An exploration of Common Crawl"
authors = [ "mateusz" ]

[tool.poetry.dependencies]
python = "^3.9"
pre-commit = "^4.0.1"
ipykernel = "^6.29.5"
aws-glue-sessions = "^1.0.7"
aws-cdk-lib = "^2.163.0"
awswrangler = "^3.9.1"
constructs = "^10.4.2"
"aws-cdk.aws-glue-alpha" = "^2.163.1-alpha.0"
"aws-cdk.aws-neptune-alpha" = "^2.163.1-alpha.0"
